fix(webview): sync the view-local buffer through wrapped apiConfiguration mutations
- E1: move the provider-release switch test into switchModeTool.spec where the switchModeTool layer is the narrowest proof

- E2: assert the state postMessage shape in the mode-lock switch test

- E3: prune invalid persisted view-state entries before applying the cap

- wrap the upsert and activate buffer-refresh calls so the view-local apiConfiguration buffer tracks the activated profile again; flat provider-settings keys no longer merge into the buffer
